chore: update svelte-jester and svelte-preprocess
