+++ package.json
@@ -121,8 +121,8 @@
     "stylelint-config-recommended-scss": "^5.0.2",
     "stylelint-scss": "^4.1.0",
     "svelte": "^3.46.4",
-    "svelte-jester": "nolanlawson/svelte-jester#auto-preprocess",
-    "svelte-preprocess": "^4.7.3",
+    "svelte-jester": "^2.3.2",
+    "svelte-preprocess": "^4.10.3",
     "svgo": "^2.8.0",
     "tachometer": "^0.5.10"
   },
