+++ rollup.config.js
@@ -3,29 +3,15 @@ import resolve from '@rollup/plugin-node-resolve'
 import replace from '@rollup/plugin-replace'
 import strip from '@rollup/plugin-strip'
 import svelte from 'rollup-plugin-svelte'
-import preprocess from 'svelte-preprocess'
 import analyze from 'rollup-plugin-analyzer'
 import { buildStyles } from './bin/buildStyles'
+import { createRequire } from 'module'
+const require = createRequire(import.meta.url)
+
+const svelteConfig = require('./svelte.config.cjs')
 
 const { NODE_ENV, DEBUG } = process.env
 const dev = NODE_ENV !== 'production'
-
-const preprocessConfig = preprocess()
-
-const origMarkup = preprocessConfig.markup
-// minify the HTML by removing extra whitespace
-// TODO: this is fragile, but it also results in a lot of bundlesize savings. let's find a better solution
-preprocessConfig.markup = async function () {
-  const res = await origMarkup.apply(this, arguments)
-
-  // remove whitespace
-  res.code = res.code.replace(/([>}])\s+([<{])/sg, '$1$2')
-
-  // remove data-testid (only used for testing-library)
-  res.code = res.code.replace(/data-testid=".*?"/g, '')
-
-  return res
-}
 
 // Build Database.test.js and Picker.js as separate modules at build times so that they are properly tree-shakeable.
 // Most of this has to happen because customElements.define() has side effects
@@ -48,7 +34,7 @@ const baseConfig = {
       compilerOptions: {
         dev
       },
-      preprocess: preprocessConfig
+      ...svelteConfig
     }),
     // make the svelte output slightly smaller
     replace({
